Made changes to allow CLASS to be linked to as an external library.
This was achieved by modifying configure.ac and Makefile.am to conditionally
link to an external libclass.so library instead of building the local copy in
the class/ subdir. The local copy will be used if the CLASSDIR env. var. is
not set. Otherwise, CLASSDIR should point to the directory in which a previously
compiled copy of libclass.so is stored.

Some small changes to the root swig file and the ccl_power.c include statements
was also needed.