Use a flatter include structure and avoid 'pragma once'
see #466.

Most public API are in ccl.h, except a few optional ones in:

ccl_emu17, ccl_halomod and ccl_lsst_specs;

on the inter-header dependency:

internal code also directly includes ccl.h to avoid tracking inter
header dependency.

if a module is included in ccl.h, then the only file that tracks this dependency is
ccl.h. otherwise a little bit of care is needed to include the headers
in ccl_emu17, ccl_halomod and ccl_lsst_specs.

I did find it a bit strange that ccl_power depends on ccl_emu17. Refactoring is possible,
but at a cost of backward compatibility and verboseness of the API.